django_ecommerce/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.7
django-crispy-forms==1.12.0
django-jalali-date==0.3.2
django-rest-passwordreset==1.2.0
djangorestframework==3.12.4
install==1.3.4
jdatetime==3.6.4
Pillow==8.3.2
psycopg2==2.9.1
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
